refactor: cleanup code and make files smaller
