+++ packages/vuelidate/src/storage.js
@@ -45,7 +45,7 @@ export default class ResultsStorage {
    * Returns the matched result if catche is valid
    * @param {String} path
    * @param {Object<NormalizedValidator>} rules
-   * @return {{$partial: boolean, $dirty: Ref<boolean>}|undefined|ValidationResult}
+   * @return {{$partial: boolean, $dirty: Ref<Boolean>, $unwatch: function}|undefined|ValidationResult}
    */
   get (path, rules) {
     const storedRuleResultPair = this.storage.get(path)
