+++ packages/vuelidate/src/utils/sortValidations.js
@@ -1,0 +1,41 @@
+import { isFunction, unwrap } from '../utils'
+
+/**
+ * Sorts a validation definition into rules, configs and nested validators.
+ * @param {Object<NormalizedValidator|Function>} validationsRaw
+ * @return {{ rules: Object<NormalizedValidator>, nestedValidators: Object, config: GlobalConfig }}
+ */
+export function sortValidations (validationsRaw = {}) {
+  const validations = unwrap(validationsRaw)
+  const validationKeys = Object.keys(validations)
+
+  const rules = {}
+  const nestedValidators = {}
+  const config = {}
+
+  validationKeys.forEach(key => {
+    const v = validations[key]
+
+    switch (true) {
+      // If it is already normalized, use it
+      case isFunction(v.$validator):
+        rules[key] = v
+        break
+      // If it is just a function, normalize it first
+      // into { $validator: <Fun> }
+      case isFunction(v):
+        rules[key] = { $validator: v }
+        break
+      // Catch $-prefixed properties as config
+      case key.startsWith('$'):
+        config[key] = v
+        break
+      // If it doesn’t match any of the above,
+      // treat as nestedValidators state property
+      default:
+        nestedValidators[key] = v
+    }
+  })
+
+  return { rules, nestedValidators, config }
+}
